utc: move UTC class definition to module
